pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perceptraderai"
version = "0.1.0"
description = "AI Trading Framework for MetaTrader5"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "MetaTrader5>=5.0.36",
    "gym>=0.26.0",
    "stable-baselines3>=2.0.0",
    "ta-lib>=0.4.24",
    "python-dotenv>=1.0.0"
]

[project.scripts]
perceptrader = "perceptrader:main"